# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema Pesquisas — API Multi-Tenant",
        "description": "Backend REST API multiempresa com RBAC. Use POST /api/v1/auth/login para obter o JWT e clique em Authorize inserindo: Bearer {token}. Super Admins podem usar o header X-Empresa-Id para escopar requisições a uma empresa.",
        "contact": {
            "email": "admin@sistema.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/usuarios": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar usuários",
                "operationId": "a25d2373e5467dc335ea0fcc231f571c",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ativo",
                                "inativo",
                                "bloqueado"
                            ]
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "gerente",
                                "usuario"
                            ]
                        }
                    },
                    {
                        "name": "filial_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Usuario"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastrar usuário",
                "description": "Valida limite max_usuarios.",
                "operationId": "14fba63bb769fe8d1cc6de821b1fa089",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha",
                                    "tipo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Maria Santos"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "gerente",
                                            "usuario"
                                        ]
                                    },
                                    "filial_id": {
                                        "type": "integer"
                                    },
                                    "telefone": {
                                        "type": "string"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/usuarios/{id}/bloquear": {
            "patch": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Bloquear usuário",
                "operationId": "361429623d921a93e81a2d39f62ebab3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bloqueado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/usuarios/{id}/ativar": {
            "patch": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Ativar usuário",
                "operationId": "4e15b5ac273ce0f19904d2c20d43e503",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ativado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Usuario": {
                "description": "Dados de um usuário",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "filial_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "nome": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@empresa.com"
                    },
                    "foto": {
                        "type": "string",
                        "nullable": true
                    },
                    "telefone": {
                        "type": "string",
                        "nullable": true
                    },
                    "tipo": {
                        "type": "string",
                        "example": "usuario",
                        "enum": [
                            "superadmin",
                            "admin",
                            "gerente",
                            "usuario"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "ativo",
                        "enum": [
                            "ativo",
                            "inativo",
                            "bloqueado"
                        ]
                    },
                    "ultimo_login": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "primeiro_acesso": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtido via POST /api/v1/auth/login. Insira: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}